requirements.txt
llama-cpp-python==0.3.4
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
gradio==5.12.0
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.1.2
pypdf==5.1.0
beautifulsoup4==4.12.3
requests==2.32.3
huggingface-hub==0.27.1
tqdm==4.67.1
numpy==1.26.4
